Make the code DRY
